Added "say" command to the minecraft utility.  This prints the given string to the in-game chat.
By utilizing the same method as is used in many other parts of the utility, this modification allows a user to print text to the in-game chat from outside the screen session.  This is done in a similar way to how other operations are preformed such as the saveon and saveoff commands: it opens the screen session and evaluates a string into the standard input stream of the minecraft server.

This is useful to add customization to many automated systems.  Without customization of the script itself, it is possible to alert users to what is going on in the server.  (Is the server being restarted? map switched? temporarily down?).  Most importantly, it lets admins administrate without having to open up a pesky screen session, (something that makes remote administration without a terminal possible).