This makes it so that "whole backup" echos to the screen to let you know that it is doing a whole backup.
This is especially usefull when running the update portion of the script,
because otherwise, it just looks like it is hanging there not doing anything.
This confused me for a while (I actually killed the process multiple times)
before realizing that it was doing a complete backup.  This pull request
simply makes it more user friendly for server admins like myself :)